[REF] runbot_gitlab: Change the way GitHub icons and texts are replaced
Instead of e.g.:
Inherit a GitHub icon and add a condition to replace the class by the
GitLab's one in case it's a GitLab repo, or leave untouched otherwise.

Now the following is performed:
- Hide elements for GitLab repos
- Add a new element that will be visible  only for GitLab repos

This new approach makes easier to support new backends in the future
(e.g. Bitbucket), because they won't cause conflicts, they would only
need to follow the same approach.

In addition, not only icons are taken into account, but also the text
"Open commit on GitHub".